Implemented custom inspector button functionality and fixes
Added simple toggle field that hides some information until it's enabled. If a value in the hidden field is set and then the field is disabled again, the value properly resets to 0. Fixed a strange issue of the user needing to double click the toggle field (first click set hidden field value to 0, second hides the field itself again). Now properly does both on one click.